Simplify define() control flow significantly
Get rid of innerDefine() and don't pass in a resolve() function to makeDefine().
Instead, deterministically check if define() has been called after control flow
resumes back to loader.js (after having eval'd the external script) to see if
the dependency was resolved by the call to define() (in the case of an AMD/UMD
module) or if it still needs to be resolved (in the case of global browser
scripts or CommonJS modules).

Note that this makes the call to define() return void (though the dependency
isn't resolved immediately - that still happens asynchronously).